requirements.txt
torch
torchvision
scikit-learn
matplotlib
seaborn
numpy
Pillow
PyYAML
kagglehub
joblib
gradio